@@ -1,9 +1,9 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.WindowsDesktop">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <UseWPF>true</UseWPF>
+    <TargetFramework>net5.0-windows</TargetFramework>
+	  <UseWPF>true</UseWPF>
     <Nullable>enable</Nullable>
     <StartupObject>Tauron.Application.Localizer.CoreProgramm</StartupObject>
     <ApplicationIcon>icon.ico</ApplicationIcon>
@@ -15,6 +15,7 @@
 
   <ItemGroup>
     <None Remove="icon.ico" />
+    <None Remove="Resource\iconfinder_Network Connection Manager_2203.png" />
   </ItemGroup>
 
   <ItemGroup>
@@ -22,26 +23,27 @@
   </ItemGroup>
 
   <ItemGroup>
+    <PackageReference Include="Dirkster.AvalonDock" Version="4.50.1" />
+    <PackageReference Include="Dirkster.AvalonDock.Themes.Metro" Version="4.50.1" />
+    <PackageReference Include="MaterialDesignExtensions" Version="3.2.0" />
     <PackageReference Include="Simple.Wpf.Terminal" Version="2.3.15" />
-    <PackageReference Include="Syncfusion.SfBusyIndicator.WPF" Version="18.1.0.57" />
-    <PackageReference Include="Syncfusion.SfGrid.WPF" Version="18.1.0.57" />
-    <PackageReference Include="Syncfusion.SfInput.WPF" Version="18.1.0.57" />
-    <PackageReference Include="Syncfusion.SfSkinManager.WPF" Version="18.1.0.57" />
-    <PackageReference Include="Syncfusion.Shared.WPF" Version="18.1.0.57" />
-    <PackageReference Include="Syncfusion.Themes.Blend.WPF" Version="18.1.0.57" />
-    <PackageReference Include="Syncfusion.Tools.WPF" Version="18.1.0.57" />
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Akka.Code.Configuration\Akka.Code.Configuration.csproj" />
+    <Page Include="App.xaml">
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\..\Shared\Tauron.Application.Wpf.SerilogViewer\Tauron.Application.Wpf.SerilogViewer.csproj" />
+    <ProjectReference Include="..\..\..\..\Shared\Tauron.Application.Wpf\Tauron.Application.Wpf.csproj" />
     <ProjectReference Include="..\Tauron.Application.Localizer.UIModels\Tauron.Application.Localizer.UIModels.csproj" />
-    <ProjectReference Include="..\Tauron.Application.Logging\Tauron.Application.Logging.csproj" />
-    <ProjectReference Include="..\Tauron.Application.Wpf.SerilogViewer\Tauron.Application.Wpf.SerilogViewer.csproj" />
-    <ProjectReference Include="..\Tauron.Application.Wpf\Tauron.Application.Wpf.csproj" />
   </ItemGroup>
 
   <ItemGroup>
     <Resource Include="icon.ico" />
+    <Resource Include="Resource\iconfinder_Network Connection Manager_2203.png" />
   </ItemGroup>
 
   <ItemGroup>
