@@ -1,7 +1,11 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
+using System.Reactive;
+using System.Reactive.Linq;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Input;
+using Tauron.Application.CommonUI;
 using Tauron.Application.Localizer.UIModels.lang;
 using Tauron.Application.Localizer.UIModels.Views;
 using Tauron.Application.Wpf;
@@ -16,7 +20,7 @@ namespace Tauron.Application.Localizer.Views
         private readonly IDialogFactory _dialogFactory;
         private readonly OpenFileMode _filemode;
         private readonly LocLocalizer _localizer;
-        private readonly TaskCompletionSource<string?> _selector = new TaskCompletionSource<string?>();
+        private readonly TaskCompletionSource<string?> _selector = new();
 
         public OpenFileDialogView(IDialogFactory dialogFactory, LocLocalizer localizer, OpenFileMode filemode)
         {
@@ -26,23 +30,21 @@ namespace Tauron.Application.Localizer.Views
             InitializeComponent();
 
             if (filemode == OpenFileMode.OpenNewFile)
-                Title += _localizer.OpenFileDialogViewHeaderNewPrefix;
+                Title.Text += _localizer.OpenFileDialogViewHeaderNewPrefix;
         }
 
-        public Task<string?> Init(string? initalData)
-        {
-            return _selector.Task;
-        }
+        public Task<string?> Init(Unit initalData) => _selector.Task;
 
         private void Search_OnClick(object sender, RoutedEventArgs e)
         {
             var result = _dialogFactory.ShowOpenFileDialog(null, _filemode == OpenFileMode.OpenExistingFile, "transp", true, _localizer.OpenFileDialogViewDialogFilter, false,
                 _localizer.OpenFileDialogViewDialogTitle,
-                true, true, out var ok);
-
-            if (ok != true) return;
-
-            PART_Path.Text = result.FirstOrDefault();
+                true, true)
+                                       .NotNull()
+                                       .Select(s => s.FirstOrDefault())
+                                       .NotNull()
+                                       .ObserveOnDispatcher()
+                                       .Subscribe(s => PART_Path.Text = s);
         }
 
         private void Ready_OnClick(object sender, RoutedEventArgs e)
