@@ -1,4 +1,4 @@
 ﻿namespace Tauron.Application.Localizer.DataModel.Processing
 {
-    public sealed record SaveProject(string OperationId, ProjectFile projectFile) : Operation(OperationId);
+    public sealed record SaveProject(string OperationId, ProjectFile ProjectFile) : Operation(OperationId);
 }
