@@ -16,7 +16,7 @@ namespace Tauron.Application.Localizer.UIModels
 
         public ImmutableList<string> RenctFiles
         {
-            get { return _renctFiles ??= ImmutableList<string>.Empty.AddRange(GetValue(s => s.Split(new[] {';'}, StringSplitOptions.RemoveEmptyEntries), Array.Empty<string>())); }
+            get { return _renctFiles ??= ImmutableList<string>.Empty.AddRange(GetValue(s => s.Split(new[] {';'}, StringSplitOptions.RemoveEmptyEntries), Array.Empty<string>())!); }
             set
             {
                 _renctFiles = value;
