@@ -2,13 +2,5 @@
 
 namespace Tauron.Application.Localizer.UIModels
 {
-    public sealed class InitProjectViewModel
-    {
-        public InitProjectViewModel(Project project)
-        {
-            Project = project;
-        }
-
-        public Project Project { get; }
-    }
+    public sealed record InitProjectViewModel(Project Project);
 }
