@@ -1,12 +1,4 @@
 ﻿namespace Tauron.Application.Localizer.UIModels.Messages
 {
-    public sealed class UpdateSource
-    {
-        public UpdateSource(string name)
-        {
-            Name = name;
-        }
-
-        public string Name { get; }
-    }
+    public sealed record UpdateSource(string Name);
 }
