@@ -1,15 +1,21 @@
 ﻿using System;
+using System.Collections.ObjectModel;
 using System.Linq;
 using System.Windows;
 using System.Windows.Input;
+using DynamicData;
+using JetBrains.Annotations;
+using Tauron.Application.CommonUI;
+using Tauron.Application.CommonUI.Commands;
 using Tauron.Application.Localizer.DataModel;
-using Tauron.Application.Wpf;
-using Tauron.Application.Wpf.Commands;
 
 namespace Tauron.Application.Localizer.UIModels
 {
-    public sealed class ProjectEntryModel : ObservableObject
+    [PublicAPI]
+    public sealed class ProjectEntryModel : ObservableObject, IDisposable
     {
+        private readonly SourceList<ProjectLangEntry> _entries = new();
+        private readonly IDisposable _connection;
         private readonly string _projectName;
         private readonly Action<(string ProjectName, string EntryName, ActiveLanguage Lang, string Content)> _updater;
 
@@ -19,35 +25,32 @@ namespace Tauron.Application.Localizer.UIModels
             _updater = updater;
             _projectName = project.ProjectName;
             EntryName = target.Key;
-            Entries = new UIObservableCollection<ProjectLangEntry>();
+
+           _connection = _entries.Connect().ObserveOnDispatcher().Bind(out var entrys).Subscribe();
+           Entries = entrys;
+
             RemoveCommand = new SimpleCommand(() => remove((_projectName, EntryName)));
             CopyCommand = new SimpleCommand(() => Clipboard.SetText(EntryName));
 
             foreach (var language in project.ActiveLanguages)
             {
-                Entries.Add(target.Values.TryGetValue(language, out var content)
-                    ? new ProjectLangEntry(EntryChanged, language, content)
-                    : new ProjectLangEntry(EntryChanged, language, string.Empty));
+                _entries.Add(target.Values.TryGetValue(language, out var content)
+                                 ? new ProjectLangEntry(EntryChanged, language, content)
+                                 : new ProjectLangEntry(EntryChanged, language, string.Empty));
             }
         }
 
         public string EntryName { get; }
 
-        public UIObservableCollection<ProjectLangEntry> Entries { get; }
+        public ReadOnlyObservableCollection<ProjectLangEntry> Entries { get; }
 
         public ICommand CopyCommand { get; }
 
         public ICommand RemoveCommand { get; }
 
-        private void EntryChanged(string content, ActiveLanguage language)
-        {
-            _updater((_projectName, EntryName, language, content));
-        }
+        private void EntryChanged(string content, ActiveLanguage language) => _updater((_projectName, EntryName, language, content));
 
-        public void AddLanguage(ActiveLanguage lang)
-        {
-            Entries.Add(new ProjectLangEntry(EntryChanged, lang, string.Empty));
-        }
+        public void AddLanguage(ActiveLanguage lang) => _entries.Add(new ProjectLangEntry(EntryChanged, lang, string.Empty));
 
         public void Update(LocEntry entry)
         {
@@ -55,10 +58,16 @@ namespace Tauron.Application.Localizer.UIModels
             {
                 var ent = Entries.FirstOrDefault(f => f.Language == activeLanguage);
                 if (ent == null)
-                    Entries.Add(new ProjectLangEntry(EntryChanged, activeLanguage, content));
+                    _entries.Add(new ProjectLangEntry(EntryChanged, activeLanguage, content));
                 else
                     ent.UpdateContent(content);
             }
+        }
+
+        public void Dispose()
+        {
+            _entries.Dispose();
+            _connection.Dispose();
         }
     }
 }
