@@ -1,15 +1,4 @@
 ﻿namespace Tauron.Application.Localizer.UIModels
 {
-    public sealed class ProjectViewLanguageModel
-    {
-        public ProjectViewLanguageModel(string name, bool isEnabled)
-        {
-            Name = name;
-            IsEnabled = isEnabled;
-        }
-
-        public string Name { get; }
-
-        public bool IsEnabled { get; }
-    }
+    public sealed record ProjectViewLanguageModel(string Name, bool IsEnabled);
 }
