@@ -2,13 +2,5 @@
 
 namespace Tauron.Application.Localizer.UIModels.Views
 {
-    public sealed class AddLanguageDialogResult
-    {
-        public AddLanguageDialogResult(CultureInfo cultureInfo)
-        {
-            CultureInfo = cultureInfo;
-        }
-
-        public CultureInfo CultureInfo { get; }
-    }
+    public sealed record AddLanguageDialogResult(CultureInfo CultureInfo);
 }
