@@ -1,8 +1,9 @@
-﻿using Tauron.Application.Wpf.Dialogs;
+﻿using System.Reactive;
+using Tauron.Application.CommonUI.Dialogs;
 
 namespace Tauron.Application.Localizer.UIModels.Views
 {
-    public interface IOpenFileDialog : IBaseDialog<string?, string?>
+    public interface IOpenFileDialog : IBaseDialog<string?, Unit>
     {
     }
 }
