@@ -1,12 +1,4 @@
 ﻿namespace Tauron.Application.Localizer.UIModels.Views
 {
-    public sealed class NewProjectDialogResult
-    {
-        public NewProjectDialogResult(string name)
-        {
-            Name = name;
-        }
-
-        public string Name { get; }
-    }
+    public sealed record NewProjectDialogResult(string Name);
 }
