@@ -5,10 +5,7 @@ namespace Tauron.Application.Localizer.Core.Converter
 {
     public sealed class OperationBadgetConverter : ValueConverterFactoryBase
     {
-        protected override IValueConverter Create()
-        {
-            return CreateCommonConverter<int, object?>(ConvertInt);
-        }
+        protected override IValueConverter Create() => CreateCommonConverter<int, object?>(ConvertInt);
 
         private static object? ConvertInt(int arg)
         {
