@@ -1,5 +1,6 @@
 ﻿using Akka.Actor;
 using Autofac;
+using Tauron.Application.CommonUI;
 using Tauron.Application.Localizer.Core.UI;
 using Tauron.Application.Localizer.DataModel.Workspace;
 using Tauron.Application.Localizer.UIModels;
@@ -7,8 +8,6 @@ using Tauron.Application.Localizer.UIModels.lang;
 using Tauron.Application.Localizer.UIModels.Services;
 using Tauron.Application.Localizer.UIModels.Views;
 using Tauron.Application.Localizer.Views;
-using Tauron.Application.Wpf;
-using Tauron.Application.Wpf.Dialogs;
 
 namespace Tauron.Application.Localizer
 {
@@ -31,7 +30,7 @@ namespace Tauron.Application.Localizer
             builder.RegisterType<LocLocalizer>().AsSelf();
             builder.Register(cc => new ProjectFileWorkspace(cc.Resolve<ActorSystem>())).AsSelf().SingleInstance();
 
-            builder.RegisterInstance(DialogCoordinator.Instance).As<IDialogCoordinator>();
+            //builder.RegisterInstance(DialogCoordinator.Instance).As<IDialogCoordinator>();
         }
     }
 }
