@@ -1,20 +1,22 @@
 ﻿using System;
 using System.ComponentModel;
 using System.Diagnostics;
+using System.IO;
 using System.Threading.Tasks;
 using System.Windows;
+using System.Windows.Input;
+using System.Xml.Serialization;
 using Akka.Actor;
-using Serilog;
-using Syncfusion.SfSkinManager;
+using AvalonDock.Layout;
+using MaterialDesignThemes.Wpf;
+using Tauron.Application.CommonUI;
+using Tauron.Application.CommonUI.AppCore;
+using Tauron.Application.CommonUI.Dialogs;
 using Tauron.Application.Localizer.DataModel.Processing;
 using Tauron.Application.Localizer.DataModel.Workspace;
 using Tauron.Application.Localizer.UIModels;
 using Tauron.Application.Localizer.UIModels.lang;
 using Tauron.Application.Localizer.UIModels.Services;
-using Tauron.Application.Wpf;
-using Tauron.Application.Wpf.AppCore;
-using Tauron.Application.Wpf.Dialogs;
-using Window = System.Windows.Window;
 
 namespace Tauron.Application.Localizer
 {
@@ -35,41 +37,36 @@ namespace Tauron.Application.Localizer
             _mainWindowCoordinator = mainWindowCoordinator;
             _coordinator = coordinator;
             _workspace = workspace;
+
             InitializeComponent();
+
+            var diag = ((IDialogCoordinatorUIEvents) _coordinator);
+
+            diag.ShowDialogEvent += o => this.ShowDialog(o);
+            diag.HideDialogEvent += () => DialogHost.IsOpen = false;
 
             _mainWindowCoordinator.TitleChanged += () => Dispatcher.BeginInvoke(new Action(MainWindowCoordinatorOnTitleChanged));
             _mainWindowCoordinator.IsBusyChanged += IsBusyChanged;
 
             Closing += OnClosing;
-            Closed += async (sender, args) =>
+            Closed += async (_, _) =>
             {
                 Shutdown?.Invoke(this, EventArgs.Empty);
 
                 await Task.Delay(TimeSpan.FromSeconds(60));
                 Process.GetCurrentProcess().Kill(false);
             };
-
-            _coordinator.OnWindowConstructed
-                += window =>
-                   {
-                       SfSkinManager.SetVisualStyle(window, VisualStyles.Blend);
-                   };
         }
-
-        public Window Window => this;
-
+        
         public event EventHandler? Shutdown;
 
-        private void IsBusyChanged()
-        {
-            Dispatcher.Invoke(() => SfBusyIndicator.IsBusy = _mainWindowCoordinator.IsBusy);
-        }
+        private void IsBusyChanged() => Dispatcher.Invoke(() => BusyIndicator.IsBusy = _mainWindowCoordinator.IsBusy);
 
         private void OnClosing(object sender, CancelEventArgs e)
         {
             if (_mainWindowCoordinator.Saved) return;
 
-            if (_coordinator.ShowModalMessageWindow(_localizer.CommonWarnig, _localizer.MainWindowCloseWarning) == false)
+            if (_coordinator.ShowModalMessageWindow(_localizer.CommonWarnig, _localizer.MainWindowCloseWarning).Result == false)
             {
                 e.Cancel = true;
                 return;
@@ -79,25 +76,24 @@ namespace Tauron.Application.Localizer
                 _workspace.ProjectFile.Operator.Tell(ForceSave.Seal(_workspace.ProjectFile), ActorRefs.NoSender);
         }
 
-        private void MainWindowCoordinatorOnTitleChanged()
-        {
-            Title = _localizer.MainWindowTitle + " - " + _mainWindowCoordinator.TitlePostfix;
-        }
+        private void MainWindowCoordinatorOnTitleChanged() => Title = _localizer.MainWindowTitle + " - " + _mainWindowCoordinator.TitlePostfix;
 
         private void FrameworkElement_OnLoaded(object sender, RoutedEventArgs e)
         {
-            try
-            {
-                DockingManager.LoadDockState();
-            }
-            catch (Exception exception)
-            {
-                Log.ForContext<MainWindow>().Error(exception, "Error on Load Dock State");
-                DockingManager.ResetState();
-            }
-        }
+            //var ser = new XmlSerializer(typeof(LayoutRoot));
+            //var testWriter = new StringWriter();
+            //ser.Serialize(testWriter, DockingManager.Layout);
+            //var resr = testWriter.ToString();
 
-        private void DockReset(object sender, RoutedEventArgs e) 
-            => DockingManager.ResetState();
+            //try
+            //{
+            //    DockingManager.LoadDockState();
+            //}
+            //catch (Exception exception)
+            //{
+            //    Log.ForContext<MainWindow>().Error(exception, "Error on Load Dock State");
+            //    DockingManager.ResetState();
+            //}
+        }
     }
 }
