@@ -40,7 +40,7 @@ namespace Tauron.Application.Workshop
                 return eventSource(_mutator);
             }
 
-            public void With(Func<TMutator, Action<TRecieve>> run) 
+            public IDisposable With(Func<TMutator, Action<TRecieve>> run) 
                 => _selector.Subscribe(run(_mutator));
         }
 
